Add a contexts module
This contains a command to get all the child contexts of a given
context, or to get all top-level contexts. It also provides events to
monitor when contexts are created or destroyed.

This is the first module to contain any events, so some infrastructure
for emitting events is also added.